requirements.txt
colorama==0.3.7
dj-database-url==0.4.1
Django==1.9.6
django-grappelli==2.8.1
django-localflavor==1.3
django-simple-address==0.1
Pillow==3.2.0
psycopg2==2.6.1
qrcode==5.2.2
qrplatba==0.3.4
reportlab==3.3.0
six==1.10.0
sorl-thumbnail==12.3
sqlparse==0.1.19

